sqlite3 doesn't accept floats so amount begins in cents

+++ db/migrate/20190410022658_create_transactions.rb
@@ -1,9 +1,11 @@
 class CreateTransactions < ActiveRecord::Migration[4.2]
   def change
     create_table :transactions do |t|
-      t.float :amount
+      t.integer :amount
       t.string :version
       t.integer :user_id
     end
   end
 end
+
+#sqlite3 doesn't accept floats so amount begins in cents
